May, 5th commit
shooter shutdown mode controller added: shutdown mode controls the PID value applied to the motors; when under work mode, shooter PID is set to some positive real value; and when under shutdown mode, the PID value is set to 0. This is designed to realize the after-shooting-0-speed function -> which is when the shooter finishes shooting, the speed of it is not directly set to 0, instead the power supply to the motor is "cut off", so the wheel will turn by inertia, making re-speeding faster and saves electricity.